proyectos/Django/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
cachetools==5.5.1
certifi==2024.12.14
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
coreschema==0.0.4
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.5.0
google-api-core==2.24.1
googleapis-common-protos==1.66.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
openai==0.28.0
packaging==24.2
pathlib==1.0.1
pillow==11.1.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==8.0.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
yarl==1.18.3
